Make week05 compile
Week05 wouldn't compile because the Plutus target (in the cabal.project
file) didn't exist. Updating this target to a commit that exists showed
lots of other compilation problems in the code like:

1. the endpoints were improperly configured
2. txInfoForge no longer exists and was replaced by txInfoMint
3. the mint function for homework 2 needed to unit parameeter to compile

This commit fixes all these issues and now the homework properly
compiles.

This commit also adds a compilation flag to the .cabal file to fix
issues with the language server.